core/mcp_hub/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
mcp==1.2.1
pydantic==2.3.0
redis==5.0.1
requests==2.31.0
python-dotenv==1.0.0
